+++ baseDAO.cabal
@@ -9,7 +9,7 @@ version:        0.1.0.0
 description:    Please see the README on GitHub at <https://github.com/tqtezos/baseDAO#readme>
 author:         Serokell, Tocqueville Group
 maintainer:     Serokell, Tocqueville Group
-copyright:      2020 TQ Tezos
+copyright:      2021 TQ Tezos
 license:        MIT
 license-file:   LICENSE
 build-type:     Simple
@@ -23,16 +23,6 @@ source-repository head
 library
   exposed-modules:
       BaseDAO.CLI
-      BaseDAO.ShareTest.Common
-      BaseDAO.ShareTest.FA2
-      BaseDAO.ShareTest.Management
-      BaseDAO.ShareTest.OffChainViews
-      BaseDAO.ShareTest.Proposal
-      BaseDAO.ShareTest.Proposal.Bounds
-      BaseDAO.ShareTest.Proposal.Config
-      BaseDAO.ShareTest.Proposal.Proposal
-      BaseDAO.ShareTest.Proposal.Vote
-      BaseDAO.ShareTest.Token
       Lorentz.Contracts.BaseDAO
       Lorentz.Contracts.BaseDAO.Common.Types
       Lorentz.Contracts.BaseDAO.Doc
@@ -60,10 +50,8 @@ library
   default-extensions: AllowAmbiguousTypes BangPatterns BlockArguments ConstraintKinds DataKinds DefaultSignatures DeriveAnyClass DeriveDataTypeable DeriveFoldable DeriveFunctor DeriveGeneric DeriveTraversable DerivingStrategies DerivingVia EmptyCase ExistentialQuantification FlexibleContexts FlexibleInstances GADTs GeneralizedNewtypeDeriving LambdaCase MultiParamTypeClasses MultiWayIf NamedFieldPuns NegativeLiterals NoImplicitPrelude NumDecimals OverloadedLabels OverloadedStrings PatternSynonyms PolyKinds QuasiQuotes RankNTypes RecordWildCards RecursiveDo ScopedTypeVariables StandaloneDeriving StrictData TemplateHaskell TupleSections TypeApplications TypeFamilies TypeOperators UndecidableInstances UndecidableSuperClasses ViewPatterns RebindableSyntax
   ghc-options: -Weverything -Wno-missing-exported-signatures -Wno-missing-import-lists -Wno-missed-specialisations -Wno-all-missed-specialisations -Wno-unsafe -Wno-safe -Wno-missing-local-signatures -Wno-monomorphism-restriction -Wno-implicit-prelude -Wno-unused-packages -Wno-prepositive-qualified-module -Wno-missing-safe-haskell-mode -Wno-unused-do-bind
   build-depends:
-      HUnit
-    , aeson-pretty
+      aeson-pretty
     , base-noprelude >=4.7 && <5
-    , cleveland
     , containers
     , fmt
     , lens
@@ -74,8 +62,6 @@ library
     , morley-prelude
     , named
     , optparse-applicative
-    , tasty
-    , tasty-hunit-compat
     , universum
   default-language: Haskell2010
 
@@ -100,11 +86,16 @@ test-suite baseDAO-test
   type: exitcode-stdio-1.0
   main-is: Main.hs
   other-modules:
+      Test.BaseDAO.Common
       Test.BaseDAO.Doc
+      Test.BaseDAO.FA2
       Test.BaseDAO.Limits
       Test.BaseDAO.Management
       Test.BaseDAO.OffChainViews
       Test.BaseDAO.Proposal
+      Test.BaseDAO.Proposal.Config
+      Test.BaseDAO.Proposal.Proposal
+      Test.BaseDAO.Proposal.Vote
       Test.BaseDAO.Token
       Test.BaseDAO.Token.FA2
       Test.Common
@@ -127,9 +118,11 @@ test-suite baseDAO-test
     , cleveland
     , containers
     , fmt
+    , lens
     , lorentz
     , morley
     , morley-ledgers
+    , morley-metadata
     , morley-prelude
     , named
     , o-clock
