+++ test/Test/Integrational/Common.hs
@@ -1,4 +1,4 @@
--- SPDX-FileCopyrightText: 2020 TQ Tezos
+-- SPDX-FileCopyrightText: 2021 TQ Tezos
 -- SPDX-License-Identifier: LicenseRef-MIT-TQ
 
 module Test.Integrational.Common
@@ -15,7 +15,7 @@ import Lorentz.Test
 import Named (defaults, (!))
 import Util.Named ((.!))
 
-import BaseDAO.ShareTest.Common (totalSupplyFromLedger)
+import Test.BaseDAO.Common (totalSupplyFromLedger)
 import qualified Lorentz.Contracts.BaseDAO as DAO
 import Lorentz.Contracts.BaseDAO.Types
