+++ test/Test/Integrational/RegistryDAO.hs
@@ -1,4 +1,4 @@
--- SPDX-FileCopyrightText: 2020 TQ Tezos
+-- SPDX-FileCopyrightText: 2021 TQ Tezos
 -- SPDX-License-Identifier: LicenseRef-MIT-TQ
 
 {-# OPTIONS_GHC -Wno-deprecations #-}
